[TASK] FAIRLUM-21: Add quality tooling

Closes FAIRLUM-21

See merge request fairway/celum-openapi-rest-client!3

+++ lib/HeaderSelector.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * HeaderSelector
  * PHP version 8.1
@@ -53,7 +54,7 @@ class HeaderSelector
         }
 
         if (!$isMultipart) {
-            if($contentType === '') {
+            if ($contentType === '') {
                 $contentType = 'application/json';
             }
 
@@ -105,17 +106,18 @@ class HeaderSelector
     {
         return preg_match('~^application/(json|[\w!#$&.+-^_]+\+json)\s*(;|$)~', $searchString) === 1;
     }
-    
+
     /**
     * Select all items from a list containing a JSON mime type
     *
     * @param array $mimeList
     * @return array
     */
-    private function selectJsonMimeList(array $mimeList): array {
+    private function selectJsonMimeList(array $mimeList): array
+    {
         $jsonMimeList = [];
         foreach ($mimeList as $mime) {
-            if($this->isJsonMime($mime)) {
+            if ($this->isJsonMime($mime)) {
                 $jsonMimeList[] = $mime;
             }
         }
@@ -157,7 +159,7 @@ class HeaderSelector
 
         $hasMoreThan28Headers = count($accept) > 28;
 
-        foreach($processedHeaders as $headers) {
+        foreach ($processedHeaders as $headers) {
             if (count($headers) > 0) {
                 $acceptHeaders[] = $this->adjustWeight($headers, $currentWeight, $hasMoreThan28Headers);
             }
@@ -181,7 +183,7 @@ class HeaderSelector
         if (preg_match('/(.*);\s*q=(1(?:\.0+)?|0\.\d+)$/', $header, $outputArray) === 1) {
             $headerData = [
                 'header' => $outputArray[1],
-                'weight' => (int)($outputArray[2] * 1000),
+                'weight' => (int)((float)$outputArray[2] * 1000),
             ];
         } else {
             $headerData = [
@@ -195,11 +197,11 @@ class HeaderSelector
 
     /**
      * @param array[] $headers
-     * @param float   $currentWeight
+     * @param int     $currentWeight
      * @param bool    $hasMoreThan28Headers
      * @return string[] array of adjusted "Accept" headers
      */
-    private function adjustWeight(array $headers, float &$currentWeight, bool $hasMoreThan28Headers): array
+    private function adjustWeight(array $headers, int &$currentWeight, bool $hasMoreThan28Headers): array
     {
         usort($headers, function (array $a, array $b) {
             return $b['weight'] - $a['weight'];
@@ -207,8 +209,7 @@ class HeaderSelector
 
         $acceptHeaders = [];
         foreach ($headers as $index => $header) {
-            if($index > 0 && $headers[$index - 1]['weight'] > $header['weight'])
-            {
+            if ($index > 0 && $headers[$index - 1]['weight'] > $header['weight']) {
                 $currentWeight = $this->getNextWeight($currentWeight, $hasMoreThan28Headers);
             }
 
@@ -229,7 +230,7 @@ class HeaderSelector
      */
     private function buildAcceptHeader(string $header, int $weight): string
     {
-        if($weight === 1000) {
+        if ($weight === 1000) {
             return $header;
         }
 
@@ -268,6 +269,6 @@ class HeaderSelector
             return $currentWeight - 1;
         }
 
-        return $currentWeight - 10 ** floor( log10($currentWeight - 1) );
+        return $currentWeight - (int)(10 ** floor(log10($currentWeight - 1)));
     }
 }
